New sphere and fixes for camera, phong shading
Limited vertical motion for the turntable camera. Fixed some issues with the phong shading's view position and highlights appearing on the opposite side. Sphere now consists of equal area triangles, I'm not sure if this actually looks better though. Might be a good idea to change it back. Also fixed some triangulation issues and added texture coordinates (Which are untested, but probably right)